fix: restore work-tile candidacy split and drop tile_keys shim (#3877)
Single-file merge exceeded repo.domain_package_source_file_size (611 > 500)
and repo.orders_no_part_directives blocks part files. Restore the prefilter
companion library (order_engine_slot pattern) and route tile-keys consumers
through order_suggestion_work_tile_candidacy.dart directly.

Refs #3877
