--- packages/colonizethis_orders/lib/src/orders/order_suggestion_work_tile_keys.dart
@@ -1,5 +1,0 @@
-export 'order_suggestion_work_tile_candidacy.dart'
-    show
-        getValidWorkOrderTileKeys,
-        getValidWorkOrderTileKeysWithVisibility,
-        sortedVisibleWorkTargetCandidates;
