+++ packages/colonizethis_orders/lib/src/orders/per_player_work_target_selection_cache.dart
@@ -8,7 +8,7 @@ import 'package:colonizethis_world/colonizethis_world.dart';
 import 'incremental_candidate_validator.dart';
 import 'order_resolution_context.dart';
 import 'order_suggestion_context.dart';
-import 'order_suggestion_work_tile_keys.dart';
+import 'order_suggestion_work_tile_candidacy.dart';
 
 /// Inputs for populating per-player work-target tile selection caches.
 ///
